move status kv code to condition package
